[Refactor] TitleScene UI
- [refactor] UIManager 리팩토링
- [feat] MainHUD 구현
- [refactor] UpgradeUI 리팩토링
- [refactor] ResourceManager 및 Facility 리팩토링

+++ Yandere/Assets/01.Scripts/Data/SaveData.cs
@@ -9,8 +9,8 @@ public class AccountSaveData
     public int accountLevel;
     public double totalPlayTime;
 
-    public int obsessionCrystals;
-    public int premiumCurrency;
+    public float obsessionCrystals;
+    public float premiumCurrency;
 
     public FacilitySaveData facilitySaveData;
     public SettingData settingData;
